♻️ Split test file by task and task grade (#1091)
